requirements.txt
# Core dependencies
apache-airflow[postgres,aws,crypto]==2.8.0
pandas==2.1.4
requests==2.31.0
boto3==1.34.0
snowflake-connector-python==3.6.0
snowflake-sqlalchemy==1.5.1

# Data processing
numpy==1.24.3
pyarrow==14.0.2
fastparquet==2023.10.1

# Configuration and utilities
python-decouple==3.8
pyyaml==6.0.1
click==8.1.7

# DBT
dbt-core==1.7.4
dbt-snowflake==1.7.1

# Testing and development
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0

# Monitoring and logging
structlog==23.2.0
sentry-sdk==1.39.2

# Airflow providers
apache-airflow-providers-amazon==8.13.0
apache-airflow-providers-snowflake==5.1.1
apache-airflow-providers-postgres==5.8.0